Move iterators to core package
